refactor/#85: Database 코드 분리
