+++ core/database/src/main/java/com/suwiki/core/database/database/converter/TimetableCellListConverter.kt
@@ -1,0 +1,18 @@
+package com.suwiki.core.database.database.converter
+
+import androidx.room.TypeConverter
+import com.suwiki.core.model.timetable.TimetableCell
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
+
+class TimetableCellListConverter {
+  @TypeConverter
+  fun cellListToJson(value: List<TimetableCell>): String {
+    return Json.encodeToString(value)
+  }
+
+  @TypeConverter
+  fun jsonToCellList(value: String): List<TimetableCell> {
+    return Json.decodeFromString(value)
+  }
+}
